+++ docs/make.jl
@@ -56,12 +56,14 @@ bibtex_plugin = CitationBibliography(
             "tutorials/stokes-flow.md",
             "tutorials/porous_media.md",
             "tutorials/ns_vs_diffeq.md",
+            "tutorials/reactive_surface.md",
             "tutorials/linear_shell.md",
             "tutorials/dg_heat_equation.md",
         ],
         "Topic guides" => [
             "Topic guide overview" => "topics/index.md",
             "topics/fe_intro.md",
+            "topics/reference_shapes.md",
             "topics/FEValues.md",
             "topics/degrees_of_freedom.md",
             "topics/assembly.md",
@@ -70,7 +72,7 @@ bibtex_plugin = CitationBibliography(
             "topics/grid.md",
             "topics/export.md"
         ],
-        "Reference" => [
+        "API reference" => [
             "Reference overview" => "reference/index.md",
             "reference/quadrature.md",
             "reference/interpolations.md",
@@ -96,7 +98,7 @@ bibtex_plugin = CitationBibliography(
         #     "gallery/topology_optimization.md",
         # ],
         "devdocs/index.md",
-        "references.md",
+        "cited-literature.md",
         ],
     plugins = [
         bibtex_plugin,